build: Switch from tsc to tsdown for compilation
For faster compilation, and to easier create separate bundles (such
as the plugin in the next commit).

Done as a precursor to moving the Expo config plugin to TypeScript:
sharing one bundler for both entry points keeps build configuration
in a single place once the plugin is added.
