+++ .gitignore
@@ -132,3 +132,4 @@ dist
 *.DS_Store
 
 android/build
+plugin/build
